Add new edge arrow types
A few changes, the arrow heads are now specified in a separate slot:
EDGE-ARROWS.   THe content is an property list where the indicators
can be :head or :tail to identify the side of the edge and the value
can be an arrow type.   The arrow types defined in this commit are:
- :arrow
- :circle
The default value of EDGE-ARROWS is (:head :arrow).

Note that in this commit the slot DIRECTION is still implemented
but not used.   In later commits the idea is that DIRECTION :backward will flip
the meaining of HEAD and TAIL.

Code wise some path functions are moved into a separate file.